+++ spec/rails_helper.rb
@@ -17,6 +17,7 @@ require 'email_spec/rspec'
 require 'faker'
 
 Dir[File.expand_path('support/**/*.rb', __dir__)].each(&method(:require))
+Dir[File.expand_path('**/shared_examples/*.rb')].each(&method(:require))
 
 # Checks for pending migrations and applies them before tests are run.
 # If you are not using ActiveRecord, you can remove these lines.